float : multiply tests use hex values instead of floats
float : multiply tests use hex values instead of floats

	much better precision and its more reliable

+++ flotante/main.h
@@ -44,7 +44,7 @@ void float32_print(Float32 p);
 
 uint8_t float32_isZero(Float32 p);
 uint8_t float32_compare(Float32 a, Float32 b);
-uint8_t float32_multiply_check(float x, float y, float z);
+uint8_t float32_multiply_check(uint32_t x, uint32_t y, uint32_t z);
 uint8_t float32_divide_check(float x, float y, float z);
 
 Float32 float32_shiftL8(Float32 x);
